backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
pydantic==2.7.3
pydantic-settings==2.3.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
itsdangerous==2.2.0
httpx==0.27.0
anyio==4.4.0
pytest==8.2.2
